Hey From France! Changed the lights to use a more intelligent light manager.
It now has priorities of what to light up to, should display semi-useful
information, and will do a little light show at the end of each match!
If we could change it to use a range of color values, that would be
awesome for displaying the useful information. Also, it would be awesome
if we had a way for the robot to know if it is on the red or blue
alliance, probably some dip switches. It makes everything have more team
spirit and stuff.