Wrote the basis for the autonomous command.
All that is left now is to correct some constants (I tried to make them
as right as possible with some math) and code the methods for navigating
obstacles. Also, idea, we could make the do nothing autonomous simply a
configuration which selects an obstacle the robot cannot navigate (the
drawbridge for example).